requirements.txt
pandas
requests
sqlalchemy==1.4.47  # Specify a compatible version of SQLAlchemy
# mysql-connector-python  # Commented out since we're using SQLite
# mysql  # Commented out since we're using SQLite
# ast
apache-airflow[bash]==2.8.1
asyncio
aiohttp
pymysql